Split build graph target tests
